스프라이트에 색을 입힐 수 있도록 hit시 색변경 방법을 수정

+++ Assets/Scripts/Enemy/Bosses/Boss.cs
@@ -15,7 +15,7 @@ public abstract class Boss : NormalEnemy
 	public override void GetDamaged(int damage)
 	{
 		base.GetDamaged(damage);
-		InGameUIManager.inst.UpdateBossHelthUI(Health / maxHealth);
+		InGameUIManager.inst.UpdateBossHelthUI((float)Health / maxHealth);
 	}
 
 	public override void GetDamagedToDeath()
